Disk formatting: Get a lot smarter about how BIOS's translate physical geometry into BIOS-friendly geometry.  With these changes, can now create a bootable partition for Windows XP.

+++ src/Vmdk/HostedSparseExtentStream.cs
@@ -192,7 +192,7 @@ namespace DiscUtils.Vmdk
             long grainStartPos = Utilities.RoundUp(_fileStream.Length, _header.GrainSize * Sizes.Sector);
 
             // Copy-on-write semantics, read the bytes from parent and write them out to this extent.
-            _parentDiskStream.Position = _diskOffset + ((grain + (_header.NumGTEsPerGT * grainTable)) * _header.GrainSize * Sizes.Sector);
+            _parentDiskStream.Position = _diskOffset + ((grain + (_header.NumGTEsPerGT * (long)grainTable)) * _header.GrainSize * Sizes.Sector);
             byte[] content = Utilities.ReadFully(_parentDiskStream, (int)_header.GrainSize * Sizes.Sector);
             _fileStream.Position = grainStartPos;
             _fileStream.Write(content, 0, content.Length);
@@ -215,12 +215,12 @@ namespace DiscUtils.Vmdk
                 Utilities.WriteBytesLittleEndian(_grainTable[i], buffer, i * 4);
             }
 
-            _fileStream.Position = _globalDirectory[_currentGrainTable] * Sizes.Sector;
+            _fileStream.Position = _globalDirectory[_currentGrainTable] * (long)Sizes.Sector;
             _fileStream.Write(buffer, 0, buffer.Length);
 
             if (_redundantGlobalDirectory != null)
             {
-                _fileStream.Position = _redundantGlobalDirectory[_currentGrainTable] * Sizes.Sector;
+                _fileStream.Position = _redundantGlobalDirectory[_currentGrainTable] * (long)Sizes.Sector;
                 _fileStream.Write(buffer, 0, buffer.Length);
             }
         }
